Refactored settings view so Twitch settings is separated from the rest
