+++ public/src/js/settingsview/SettingsView.jsx
@@ -1,11 +1,12 @@
 import React, { PureComponent } from "react";
 import PropTypes from "prop-types";
-import { Link } from "react-router-dom";
+import { NavLink } from "react-router-dom";
 
 import SettingsFriendsView from "./SettingsFriendsView.jsx";
 import SettingsGeneralView from "./SettingsGeneralView.jsx";
 import SettingsIrcView from "./SettingsIrcView.jsx";
 import SettingsNicknamesView from "./SettingsNicknamesView.jsx";
+import SettingsTwitchView from "./SettingsTwitchView.jsx";
 import { SETTINGS_PAGE_NAMES } from "../constants";
 import { settingsUrl } from "../lib/routeHelpers";
 
@@ -35,6 +36,9 @@ class SettingsView extends PureComponent {
 			case "nicknames":
 				page = <SettingsNicknamesView />;
 				break;
+			case "twitch":
+				page = <SettingsTwitchView />;
+				break;
 			default:
 				page = <SettingsGeneralView />;
 		}
@@ -51,9 +55,12 @@ class SettingsView extends PureComponent {
 						<ul className="settingsview__tabs switcher" key="tabs">
 							{ items.map((item) => (
 								<li key={item}>
-									<Link className={item} to={settingsUrl(item)}>
+									<NavLink
+										exact
+										to={settingsUrl(item)}
+										activeClassName="selected">
 										{ SETTINGS_PAGE_NAMES[item] }
-									</Link>
+									</NavLink>
 								</li>
 							)) }
 						</ul>
